unblock the compile errors for CLI
